Move course editing to same place as script and lesson
