+++ apps/src/lib/levelbuilder/course-editor/CourseScriptsEditor.js
