+++ apps/src/lib/levelbuilder/script-editor/LessonGroupCard.jsx
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import {connect} from 'react-redux';
-import color from '../../util/color';
+import color from '@cdo/apps/util/color';
 import {borderRadius} from '@cdo/apps/lib/levelbuilder/constants';
 import OrderControls from '@cdo/apps/lib/levelbuilder/OrderControls';
 import LessonCard from '@cdo/apps/lib/levelbuilder/script-editor/LessonCard';
