+++ apps/test/unit/templates/courseOverview/ResourcesEditorTest.js
@@ -1,7 +1,7 @@
 import {assert} from '../../../util/deprecatedChai';
 import React from 'react';
 import {shallow, mount} from 'enzyme';
-import ResourcesEditor from '@cdo/apps/templates/courseOverview/ResourcesEditor';
+import ResourcesEditor from '@cdo/apps/lib/levelbuilder/course-editor/ResourcesEditor';
 import ResourceType from '@cdo/apps/templates/courseOverview/resourceType';
 
 const defaultProps = {
